feat: create phosphor-tile-engine library, move IAutotileSettings there
New INTERFACE library for the autotile placement engine. IAutotileSettings
moves from phosphor-engine-types to its proper home — the engine's own
library owns its settings contract.

phosphor-engine-types stays universal: IPlacementState, NavigationContext,
EngineTypes only. No engine-specific interfaces.

Library will become SHARED when engine source files are moved from
src/autotile/ in subsequent commits.
