+++ CMakeLists.txt
@@ -333,6 +333,7 @@ add_subdirectory(libs/phosphor-audio)       # audio spectrum provider (CAVA back
 add_subdirectory(libs/phosphor-surfaces)    # managed surface lifecycle (engine, keep-alive, scope gen)
 add_subdirectory(libs/phosphor-protocol)    # D-Bus wire types and service constants
 add_subdirectory(libs/phosphor-engine-api)  # unified placement engine contracts (INTERFACE)
+add_subdirectory(libs/phosphor-tile-engine) # autotile placement engine (IAutotileSettings + engine code)
 
 # Source directories
 add_subdirectory(src)
