+++ libs/phosphor-tile-engine/include/PhosphorTileEngine/IAutotileSettings.h
